requirements.txt
# Common dependencies for backend services
flask
requests
openai
python-dotenv
flask-cors
pyaudio
sounddevice
webrtcvad
